requirements.txt
python-telegram-bot
aiohttp
beautifulsoup4
motor
pymongo
python-dotenv
dnspython